backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
asyncpg==0.30.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==46.0.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.117.1
Flask==3.1.1
fonttools==4.58.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
gymnasium==1.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyzmq==26.4.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
Werkzeug==3.1.3
win32_setctime==1.2.0
